app/requirements.txt
fastapi
uvicorn
psycopg2-binary
prometheus_client
sqlalchemy